ISSUE #3933 - fix cyclical dependency
